(Untested) Float free version of PID library

+++ libs/PID_v1/PID_v1.h
@@ -15,8 +15,8 @@ class PID
   #define REVERSE  1
 
   //commonly used functions **************************************************************************
-    PID(double*, double*, double*,        // * constructor.  links the PID to the Input, Output, and 
-        double, double, double, int);     //   Setpoint.  Initial tuning parameters are also set here
+    PID(long*, long*, long*,        // * constructor.  links the PID to the Input, Output, and 
+        byte, byte, byte, byte);     //   Setpoint.  Initial tuning parameters are also set here
 	
     void SetMode(int Mode);               // * sets PID to either Manual (0) or Auto (non-0)
 
@@ -25,17 +25,17 @@ class PID
                                           //   calculation frequency can be set using SetMode
                                           //   SetSampleTime respectively
 
-    void SetOutputLimits(double, double); //clamps the output to a specific range. 0-255 by default, but
+    void SetOutputLimits(long, long); //clamps the output to a specific range. 0-255 by default, but
 										  //it's likely the user will want to change this depending on
 										  //the application
 	
 
 
   //available but not commonly used functions ********************************************************
-    void SetTunings(double, double,       // * While most users will set the tunings once in the 
-                    double);         	  //   constructor, this function gives the user the option
+    void SetTunings(byte, byte,       // * While most users will set the tunings once in the 
+                    byte);         	  //   constructor, this function gives the user the option
                                           //   of changing tunings during runtime for Adaptive control
-	void SetControllerDirection(int);	  // * Sets the Direction, or "Action" of the controller. DIRECT
+	void SetControllerDirection(byte);	  // * Sets the Direction, or "Action" of the controller. DIRECT
 										  //   means the output will increase when error is positive. REVERSE
 										  //   means the opposite.  it's very unlikely that this will be needed
 										  //   once it is set in the constructor.
@@ -45,35 +45,35 @@ class PID
 										  
 										  
   //Display functions ****************************************************************
-	double GetKp();						  // These functions query the pid for interal values.
-	double GetKi();						  //  they were created mainly for the pid front-end,
-	double GetKd();						  // where it's important to know what is actually 
+	byte GetKp();						  // These functions query the pid for interal values.
+	byte GetKi();						  //  they were created mainly for the pid front-end,
+	byte GetKd();						  // where it's important to know what is actually 
 	int GetMode();						  //  inside the PID.
 	int GetDirection();					  //
 
   private:
 	void Initialize();
 	
-	double dispKp;				// * we'll hold on to the tuning parameters in user-entered 
-	double dispKi;				//   format for display purposes
-	double dispKd;				//
+	byte dispKp;				// * we'll hold on to the tuning parameters in user-entered 
+	byte dispKi;				//   format for display purposes
+	byte dispKd;				//
     
-	double kp;                  // * (P)roportional Tuning Parameter
-    double ki;                  // * (I)ntegral Tuning Parameter
-    double kd;                  // * (D)erivative Tuning Parameter
+	byte kp;                  // * (P)roportional Tuning Parameter
+    byte ki;                  // * (I)ntegral Tuning Parameter
+    byte kd;                  // * (D)erivative Tuning Parameter
 
 	int controllerDirection;
 
-    double *myInput;              // * Pointers to the Input, Output, and Setpoint variables
-    double *myOutput;             //   This creates a hard link between the variables and the 
-    double *mySetpoint;           //   PID, freeing the user from having to constantly tell us
+    long *myInput;              // * Pointers to the Input, Output, and Setpoint variables
+    long *myOutput;             //   This creates a hard link between the variables and the 
+    long *mySetpoint;           //   PID, freeing the user from having to constantly tell us
                                   //   what these values are.  with pointers we'll just know.
 			  
 	unsigned long lastTime;
-	double ITerm, lastInput;
+	long ITerm, lastInput;
 
 	unsigned long SampleTime;
-	double outMin, outMax;
+	long outMin, outMax;
 	bool inAuto;
 };
 #endif
